% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFunctions.R
\name{L_layer_model}
\alias{L_layer_model}
\title{Main function for multilayer NN estimation. Binary choice as output layer. Hidden layers use 'relu', output layer is 'sigmoid'}
\usage{
L_layer_model(X, Y, layers_dims, learning_rate = 0.0075,
  num_iterations = 3000, print_cost = F)
}
\arguments{
\item{X}{matrix of features. Each column represents one training example}

\item{Y}{matrix (1 x m) of true responses}

\item{layers_dims}{structure of the model}

\item{learning_rate}{spcecify the learning rate for gradient descent}

\item{num_iterations}{number of iterations for gradient descent}

\item{print_cost}{boolean, whether or not to print cost}
}
\description{
Main function for multilayer NN estimation. Binary choice as output layer. Hidden layers use 'relu', output layer is 'sigmoid'
}
\examples{
\dontrun{
first_layer<-nrow(X)
model_dim<-c(first_layer,,l2,l3,...,1)
parameters<-L_layer_model(X, Y, layers_dims=model_dim)
}
}
