% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFunctions.R
\name{predict_nn_L_layer}
\alias{predict_nn_L_layer}
\title{Function creates prediction based on the data and patameter estimates}
\usage{
predict_nn_L_layer(X, parameters)
}
\arguments{
\item{X}{a matrix containing each training example as column}

\item{parameters}{a matrix of weights}
}
\description{
Function creates prediction based on the data and patameter estimates
}
