% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X}
\alias{X}
\title{X: data contains images 64x64xRGB unrolled to a single column vector for each training example
some images are images of cats and some are not}
\format{matricies, column is a training example
\describe{
\item{many_more_other_items}{some other info}
}}
\source{
Coursera, deeplearning.ai
}
\usage{
X
}
\description{
X: data contains images 64x64xRGB unrolled to a single column vector for each training example
some images are images of cats and some are not
}
\examples{
\dontrun{
X
}
}
\keyword{datasets}
