% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y}
\alias{Y}
\title{some images are images of cats and some are not
Y: 1 for cat, 0 otherwiwse}
\format{matricies, column is a training example
\describe{
\item{many_more_other_items}{some other info}
}}
\source{
Coursera, deeplearning.ai
}
\usage{
Y
}
\description{
some images are images of cats and some are not
Y: 1 for cat, 0 otherwiwse
}
\examples{
\dontrun{
Y
}
}
\keyword{datasets}
